feat: 添加libtorch安装，修复BUG

+++ InstallPackages.sh
@@ -48,10 +48,19 @@ function echocolor() {
 }
 
 ##############################################
+# FunctionName:version_** 
+##############################################
+function version_gt() { test "$(echo "$@" | tr " " "\n" | sort -V | head -n 1)" != "$1"; } #>
+function version_le() { test "$(echo "$@" | tr " " "\n" | sort -V | head -n 1)" == "$1"; } #<=
+function version_lt() { test "$(echo "$@" | tr " " "\n" | sort -rV | head -n 1)" != "$1"; } #<
+function version_ge() { test "$(echo "$@" | tr " " "\n" | sort -rV | head -n 1)" == "$1"; } #>=
+
+##############################################
 # FunctionName:UpdateSource
 ##############################################
 function UpdateSource() {
     echocolor "Update Source" "purple"
+
     if [[ -f /etc/apt/sources.list.bak ]]; then
         echocolor "sources.list.bak exists" "yellow"
     else
@@ -70,6 +79,20 @@ function UpdateSource() {
     echo "deb-src http://mirrors.tuna.tsinghua.edu.cn/ubuntu-ports/ bionic-backports main multiverse restricted universe" >>/etc/apt/sources.list
 
     echocolor "Ubuntu source has been updated to Tsinghua source." "green"
+
+    sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros-latest.list'
+
+    apt-key adv --keyserver 'hkp://keyserver.ubuntu.com:80' --recv-key C1CF6E31E6BADE8868B172B4F42ED6FBAB17C654
+    if [ $? -eq 1 ]; then
+        curl -sSL 'http://keyserver.ubuntu.com/pks/lookup?op=get&search=0xC1CF6E31E6BADE8868B172B4F42ED6FBAB17C654' | apt-key add -
+    fi
+
+    $systemPackage update
+    if [ $? -eq 1 ]; then
+        echocolor "Get apt update list failed." "red"
+    fi
+
+    echo
 }
 
 ##############################################
@@ -79,8 +102,9 @@ function InstallSimpleProgram() {
     echo
     program=$1
     echocolor "Install $program" "purple"
-    which $program
-    if [ $? -eq 1 ]; then
+
+    programName=$(dpkg -l | grep $program | awk 'NR==1 {print $2 " " $3}')
+    if [[ -z $programName ]]; then
         $systemPackage -y install $program
         if [ $? -ne 0 ]; then
             echocolor "$program install failed, try re-install..." "red"
@@ -96,6 +120,7 @@ function InstallSimpleProgram() {
     else
         echocolor "$program already installed." "yellow"
     fi
+
     echo
 }
 
@@ -105,45 +130,35 @@ function InstallSimpleProgram() {
 function InstallROS() {
     echo
     echocolor "Install ROS" "purple"
-    sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros-latest.list'
 
-    apt-key adv --keyserver 'hkp://keyserver.ubuntu.com:80' --recv-key C1CF6E31E6BADE8868B172B4F42ED6FBAB17C654
-    if [ $? -eq 1 ]; then
-        curl -sSL 'http://keyserver.ubuntu.com/pks/lookup?op=get&search=0xC1CF6E31E6BADE8868B172B4F42ED6FBAB17C654' | apt-key add -
-    fi
-
-    $systemPackage update
-
+    #ROS dependencies
     InstallSimpleProgram "python-rosdep"
     InstallSimpleProgram "python-rosinstall"
     InstallSimpleProgram "python-rosinstall-generator"
     InstallSimpleProgram "python-wstool"
     InstallSimpleProgram "python-catkin-tools"
 
-    if [ $? -eq 1 ]; then
-        echocolor "Get apt update list failed." "red"
-    fi
+    rosversion -d
+    if [ $? -eq 1 ];then
+        InstallSimpleProgram "ros-melodic-desktop-full"
 
-    $systemPackage install ros-melodic-desktop-full
-    if [ $? -eq 0 ]; then
-        echocolor "ROS Desktop Full Installed Successfully." "green"
-    else
-        echocolor "ROS Desktop Full Installed Failed." "red"
-    fi
+        echo "source /opt/ros/melodic/setup.bash" >>~/.bashrc
+        source ~/.bashrc
+        echocolor "Add ros source to .bashrc successfully." "green"
 
-    echo "source /opt/ros/melodic/setup.bash" >>~/.bashrc
-    source ~/.bashrc
-    echocolor "Add ros source to .bashrc successfully." "green"
-
-    rosdep init
-    if [ $? -eq 1 ]; then
-        wget http://packages.ros.org/ros.key -O - | apt-key add -
         rosdep init
+        if [ $? -eq 1 ]; then
+            wget http://packages.ros.org/ros.key -O - | apt-key add -
+            rosdep init
+        fi
+        rosdep update
+        if [ $? -eq 1 ]; then
+            echocolor "Rosdep init failed." "red"
+        fi
+    else
+        echocolor "ROS already installed." "yellow"
     fi
-    rosdep update
-    if [ $? -eq 1 ]; then
-        echocolor "Rosdep init failed." "red"
-    fi
+
     echo
 }
 
@@ -153,6 +168,7 @@ function InstallROS() {
 function InstallOpenCV() {
     echo
     echocolor "Install OpenCV" "purple"
+
     #OpenCV dependencies
     InstallSimpleProgram "build-essential"
     InstallSimpleProgram "libgtk2.0-dev"
@@ -163,13 +179,12 @@ function InstallOpenCV() {
     InstallSimpleProgram "libswscale-dev"
     InstallSimpleProgram "libjasper-dev"
 
-    opencv_version
-    if [ $? -ne 1 ]; then
-        echocolor "OpenCV already installed." "yellow"
-    else
-        folder = ${HOME}/src
-        mkdir -p $folder
-        cd $folder
+    #opencv_version
+    opencvVersion=$(pkg-config opencv --modversion)
+    version=3.4.14
+
+    if version_lt $opencvVersion $version; then
+        cd ${HOME}/src
 
         if [[ -f ./opencv-3.4.14.tar.gz ]]; then
             echocolor "opencv-3.4.14.tar.gz exists" "yellow"
@@ -212,6 +227,8 @@ function InstallOpenCV() {
         else
             echocolor "Install OpenCV Success, Install Address at /usr." "red"
         fi
+    else
+        echocolor "OpenCV already installed." "yellow"
     fi
 
     echo
@@ -223,44 +240,49 @@ function InstallOpenCV() {
 function InstallProtobuf() {
     echo
     echocolor "Install Protobuf" "purple"
+
     #Protobuf dependencies
     InstallSimpleProgram "autoconf"
     InstallSimpleProgram "libtool"
 
-    folder = ${HOME}/src
-    mkdir -p $folder
-    cd $folder
+    protoc --version
+    if [ $? -eq 1 ];then
+        cd ${HOME}/src
 
-    if [ ! -f protobuf-python-3.8.0.zip ]; then
-        wget https://github.com/protocolbuffers/protobuf/releases/download/v3.8.0/protobuf-python-3.8.0.zip
+        if [ ! -f protobuf-python-3.8.0.zip ]; then
+            wget https://github.com/protocolbuffers/protobuf/releases/download/v3.8.0/protobuf-python-3.8.0.zip
+        fi
+
+        if [ ! -f protoc-3.8.0-linux-aarch_64.zip ]; then
+            wget https://github.com/protocolbuffers/protobuf/releases/download/v3.8.0/protoc-3.8.0-linux-aarch_64.zip
+        fi
+
+        unzip protobuf-python-3.8.0.zip
+        unzip protoc-3.8.0-linux-aarch_64.zip -d protoc-3.8.0
+        cp protoc-3.8.0/bin/protoc /usr/local/bin/protoc
+
+        export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=cpp
+        cd protobuf-3.8.0/
+
+        ./autogen.sh
+        ./configure --prefix=/usr/local
+        make -j $(nproc)
+        make check
+        make -j $(nproc) install
+        ldconfig
+
+        pip3 uninstall -y protobuf
+        pip3 install Cython
+        cd python/
+        python3 setup.py build --cpp_implementation
+        python3 setup.py test --cpp_implementation
+        sudo python3 setup.py install --cpp_implementation
+        echocolor "Install Protobuf Success, Install Address at /usr." "purple"
+    else
+        echocolor "Protobuf already Installed." "yellow"
     fi
 
-    if [ ! -f protoc-3.8.0-linux-aarch_64.zip ]; then
-        wget https://github.com/protocolbuffers/protobuf/releases/download/v3.8.0/protoc-3.8.0-linux-aarch_64.zip
-    fi
-
-    unzip protobuf-python-3.8.0.zip
-    unzip protoc-3.8.0-linux-aarch_64.zip -d protoc-3.8.0
-    cp protoc-3.8.0/bin/protoc /usr/local/bin/protoc
-
-    export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=cpp
-    cd protobuf-3.8.0/
-
-    ./autogen.sh
-    ./configure --prefix=/usr/local
-    make -j $(nproc)
-    make check
-    make -j $(nproc) install
-    ldconfig
-
-    pip3 uninstall -y protobuf
-    pip3 install Cython
-    cd python/
-    python3 setup.py build --cpp_implementation
-    python3 setup.py test --cpp_implementation
-    sudo python3 setup.py install --cpp_implementation
-    echocolor "Install Protobuf Success, Install Address at /usr." "purple"
-    echo/
+    echo
 }
 
 ##############################################
@@ -268,19 +290,36 @@ function InstallProtobuf() {
 ##############################################
 function Installlibtorch() {
     echo
-    echocolor "Install Protobuf" "purple"
+    echocolor "Install libtorch" "purple"
+    
+    #libtorch dependencies
+    InstallSimpleProgram "python3-pip"
+    InstallSimpleProgram "libopenblas-base"
+    InstallSimpleProgram "libopenmpi-dev"
 
-    folder = ${HOME}/src
-    mkdir -p $folder
-    cd $folder
+    torchVersion=$(python3 -c "import torch;print(torch.__version__)")
+    installVersion=1.7.0
 
-    if [ ! -f libtorch-shared-with-deps-1.7.0.zip ]; then
-        wget https://download.pytorch.org/libtorch/cu102/libtorch-shared-with-deps-1.7.0.zip
+    if version_lt $torchVersion $installVersion;then
+        cd ${HOME}/src
+
+        if [ ! -f torch-1.7.0-cp36-cp36m-linux_aarch64.whl ]; then
+            wget https://nvidia.box.com/shared/static/cs3xn3td6sfgtene6jdvsxlr366m2dhq.whl -O torch-1.7.0-cp36-cp36m-linux_aarch64.whl
+        fi
+
+        pip3 install Cython
+
+        pip3 install numpy torch-1.7.0-cp36-cp36m-linux_aarch64.whl
+        if [ $? -eq 1 ]; then
+            echocolor "Install libtorch failed." "red"
+            exit
+        else
+            echocolor "Install libtorch Success, Install Address at ~/src/libtorch." "purple"
+        fi
+    else
+        echocolor "libtorch already installed." "yellow"
     fi
 
-    
-
-    echocolor "Install libtorch Success, Install Address at ~/src/libtorch." "purple"
     echo
 }
 
@@ -290,14 +329,11 @@ function Installlibtorch() {
 function InstallCmake() {
     echo
     echocolor "Install Cmake" "purple"
-    
-    version=3.20
-    build=3
+
+    version=3.20.3
     cmakeVersion=$(cmake --version | awk 'NR==1 {print $3}')
 
-    if [[ $cmakeVersion > $version.$build ]]  || [[ $cmakeVersion == $version.$build ]]; then
-        echocolor "Cmake already install." "yellow"
-    else
+    if version_lt $cmakeVersion $version; then
         apt remove cmake
         if [ $? -eq 1 ]; then
             echocolor "Remove older version of cmake failed." "red"
@@ -306,9 +342,7 @@ function InstallCmake() {
             echocolor "Remove older version of cmake succeed." "green"
         fi
 
-        folder = ${HOME}/src
-        mkdir -p $folder
-        cd $folder
+        cd ${HOME}/src
 
         if [ ! -f cmake-$version.$build.tar.gz ]; then
             wget https://cmake.org/files/v$version/cmake-$version.$build.tar.gz
@@ -340,12 +374,14 @@ function InstallCmake() {
         else
             echocolor "Install cmake succeed." "green"
         fi
-
+        
         ln -s /usr/local/bin/cmake /usr/bin/cmake
-
-        echocolor "Install Cmake Success." "purple"
-        echo
+        echocolor "Install Cmake Success." "purple"        
+    else
+        echocolor "Cmake already install." "yellow"
     fi
+    
+    echo
 }
 
 ##############################################
@@ -354,6 +390,7 @@ function InstallCmake() {
 function GetSuperUserPermission() {
     echo
     echocolor "Get SuperUser Permission" "purple"
+
     sudo su
     if [ $? -eq 1 ]; then
         echocolor "Get SuperUser Permission failed." "red"
@@ -361,6 +398,7 @@ function GetSuperUserPermission() {
     else
         echocolor "Get SuperUser Permission succeed." "green"
     fi
+
     echo
 }
 
@@ -369,10 +407,11 @@ function GetSuperUserPermission() {
 ##############################################
 function GetCurrentVersion() {
     shellVersion="0.0.1"
-    echocolor "Install Package Shell Version: $shellVersion" "cyan"
+    echocolor "Install Packages Shell Version: $shellVersion" "cyan"
 
     echo
     echocolor "Get Current Version" "purple"
+
     source /etc/os-release
     RELEASE=$ID
     VERSION=$VERSION_ID
@@ -388,8 +427,8 @@ function GetCurrentVersion() {
     fi
 
     systempwd="/etc/systemd/system/"
-    jetpackVersion=$(cat /etc/nv_tegra_release)
-    tensorRTVersion=$(dpkg -l | grep TensorRT | head -1)
+    jetpackVersion=$(cat /etc/nv_tegra_release | awk '{print $2$3$4$5}')
+    tensorRTVersion=$(dpkg -l | grep TensorRT | head -1 | awk '{print $3}')
     python2Version=$(python -V 2>&1 | awk '{print $2}')
     python3Version=$(python3 -V 2>&1 | awk '{print $2}')
 
@@ -407,12 +446,15 @@ function GetCurrentVersion() {
 function InstallRequirements() {
     echo
     echocolor "Install Requirements" "purple"
+
     echocolor "Do you want to start install requirements? (Y/N)" "blue"
     read startDependencies
     if [ $startRequirements != "Y" ] && [ $startRequirements != "y" ]; then
         echocolor "Install requirements failed!" "red"
         exit 0
     fi
+
+    mkdir -p ${HOME}/src
 
     UpdateSource
 
@@ -431,6 +473,7 @@ function InstallRequirements() {
     InstallCmake
 
     #InstallProtobuf &
+    InstallProtobuf
     InstallROS
     InstallOpenCV
     Installlibtorch
